WIP - finalizado endpoint para informar resultado

+++ Backend/HealthLabApi/src/main/java/com/porto/HealthLabApi/domain/layout/DTO/RequestCadastrarLayoutCampos.java
@@ -9,7 +9,7 @@ import jakarta.validation.constraints.Size;
 
 public record RequestCadastrarLayoutCampos(
     @NotNull
-    Integer codigoCampo,
+    Long codigoCampo,
     @NotNull
     @Enumerated
     TipoCampo tipoCampo,
